@@ -27,10 +27,9 @@ class AuthorController extends ModuleController
         'editInModal' => false,
     ];
 
-    protected $indexColumns = [
+    protected ?array $indexColumns = [
         'avatar' => [
             'thumb' => true,
-
             'variant' => [
                 'role' => 'featured',
                 'crop' => 'default',
