@@ -61,7 +61,7 @@ class GroupController extends ModuleController
         ];
     }
 
-    protected $indexColumns = [
+    protected ?array $indexColumns = [
         'name' => [
             'title' => 'Name',
             'field' => 'name',
