@@ -195,7 +195,7 @@ if (! function_exists('fix_directory_separator')) {
 }
 
 if (!function_exists('twillModel')) {
-    function twillModel($model)
+    function twillModel($model): string
     {
         return config("twill.models.$model")
             ?? abort(500, "helpers/twillModel: '$model' model is not configured");
