@@ -2,7 +2,9 @@
 
 namespace A17\Twill\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model as BaseModel;
+use Illuminate\Database\Eloquent\Relations\MorphTo;
 
 class Feature extends BaseModel
 {
@@ -14,19 +16,19 @@ class Feature extends BaseModel
         'starred',
     ];
 
-    public function featured()
+    public function featured(): MorphTo
     {
         return $this->morphTo();
     }
 
-    public function scopeForBucket($query, $bucketKey)
+    public function scopeForBucket(Builder $query, string $bucketKey): Builder
     {
         return $query->where('bucket_key', $bucketKey)->get()->map(function ($feature) {
             return $feature->featured;
         })->filter();
     }
 
-    public function getTable()
+    public function getTable(): string
     {
         return config('twill.features_table', 'twill_features');
     }
