@@ -109,7 +109,7 @@ abstract class TestCase extends OrchestraTestCase
     protected function makeAllTwillPaths(): void
     {
         collect($this->paths)->each(function ($directory) {
-            if (! file_exists($directory = twill_path($directory))) {
+            if (!file_exists($directory = twill_path($directory))) {
                 $this->files->makeDirectory($directory, 0755, true);
             }
         });
@@ -353,9 +353,9 @@ abstract class TestCase extends OrchestraTestCase
     public function superAdmin($force = false)
     {
         return $this->superAdmin =
-        ! $this->superAdmin || $force
-        ? $this->makeNewSuperAdmin()
-        : $this->superAdmin;
+            !$this->superAdmin || $force
+                ? $this->makeNewSuperAdmin()
+                : $this->superAdmin;
     }
 
     /**
@@ -389,7 +389,10 @@ abstract class TestCase extends OrchestraTestCase
 
         if ($this->example) {
             $this->artisan('twill:install ' . $this->example)
-                ->expectsConfirmation('Are you sure to install this preset? This can overwrite your models, config and routes.', 'yes')
+                ->expectsConfirmation(
+                    'Are you sure to install this preset? This can overwrite your models, config and routes.',
+                    'yes'
+                )
                 ->expectsQuestion('Enter an email', $this->superAdmin()->email)
                 ->expectsQuestion('Enter a password', $this->superAdmin()->password)
                 ->expectsQuestion(
@@ -444,7 +447,7 @@ abstract class TestCase extends OrchestraTestCase
 
         return collect($routes)->filter(function ($route) {
             return Str::startsWith($route->action['uses'], 'A17\Twill') ||
-            Str::startsWith($route->action['uses'], 'App\\');
+                Str::startsWith($route->action['uses'], 'App\\');
         });
     }
 
@@ -573,7 +576,7 @@ abstract class TestCase extends OrchestraTestCase
 
                 $destination = $this->makeFileName($destination, $source);
 
-                if (! $this->files->exists($directory = dirname($destination))) {
+                if (!$this->files->exists($directory = dirname($destination))) {
                     $this->files->makeDirectory($directory, 0755, true);
                 }
 
@@ -595,11 +598,11 @@ abstract class TestCase extends OrchestraTestCase
                 File::deleteDirectory($file);
             }
 
-            if (! is_dir($file) && file_exists($file)) {
+            if (!is_dir($file) && file_exists($file)) {
                 unlink($file);
             }
 
-            if (! Str::endsWith($file, '.php')) {
+            if (!Str::endsWith($file, '.php')) {
                 File::makeDirectory($file, 0755, true);
             }
         });
@@ -642,7 +645,7 @@ abstract class TestCase extends OrchestraTestCase
 
         $extension = pathinfo($file, PATHINFO_EXTENSION);
 
-        if (filled($source) && ! Str::endsWith($file, ".{$extension}")) {
+        if (filled($source) && !Str::endsWith($file, ".{$extension}")) {
             $file = $file . basename($source);
         }
 
@@ -690,7 +693,7 @@ abstract class TestCase extends OrchestraTestCase
      */
     public function skipOnTravis()
     {
-        if (! is_null(env('TRAVIS_PHP_VERSION'))) {
+        if (!is_null(env('TRAVIS_PHP_VERSION'))) {
             $this->markTestSkipped('This test cannot be executed on Travis');
         }
     }
